PAE-1339: split validation helpers into own module
